Move NotificationTimeSpan to isolated file
