+++ frontend/js/models/NotificatonTimeSpan.ts
@@ -1,0 +1,9 @@
+// TODO: Add method isZero() and parse().
+export class NotificationTimeSpan {
+  constructor(
+    public fromHour: number,
+    public fromMinute: number,
+    public toHour: number,
+    public toMinute: number,
+  ) {}
+}
